Moved expiry list management into separate plugins -- this allows for different expiry profiles to be developed and used, rather than forcing our own ideas on how flow expiry should work onto people.
The external API is unchanged, but there are a number of new config options to handle choosing and configuring an expiry profile so I've bumped the minor version to indicate that this is not an entirely incremental change.

Added a plugin for doing simple "fixed period of inactivity" flow expiry.

Updated the example to use the new config option to specify the expiry plugin.